gemspec: switch summary and description
The summary variable in the gemspec is intended to be a small bit of
text that presents the purpose of the gem in a short "at a glance"
format

The description variable is a longer characterization of the gem.

Prior to this commit, the two settings were reversed in the gemspec.
Flip them here.