Fix arity check for keyword arguments
Previously keyword arguments were ignored, which prevented them from
being passed in. This at least allows them to be passed in by treating
them as a single optional or required argument. However, it doesn't
yet check that a hash is passed in for the keyword arguments or that
the hash has the expected keys.